--- tests/http/__init__.py
@@ -1,5 +1,0 @@
-from fastapi.testclient import TestClient
-
-from app.main import app
-
-client = TestClient(app)
